Split interface into parts
